+++ src/app/datasets/dataset-view/dataset-view.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { ChangeDetectorRef, Component, ElementRef, HostListener, ViewChild } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Project } from 'src/app/models/Project';
 import { ShowVocUrlParams } from 'src/app/models/ShowVoc';
@@ -15,13 +15,17 @@ import { SVContext } from 'src/app/utils/SVContext';
 })
 export class DatasetViewComponent {
 
+    @ViewChild("navtabs") navtabsEl: ElementRef;
+
     project: Project;
 
     hideDatasetName: boolean;
 
     isUserAuthorized: boolean;
 
-    constructor(private router: Router, private activatedRoute: ActivatedRoute) { }
+    titleMaxWidth: number = 0;
+
+    constructor(private router: Router, private activatedRoute: ActivatedRoute, private cdRef: ChangeDetectorRef) { }
 
     ngOnInit() {
         this.project = SVContext.getWorkingProject();
@@ -31,6 +35,10 @@ export class DatasetViewComponent {
                 this.hideDatasetName = params[ShowVocUrlParams.hideDatasetName] == "true";
             }
         );
+    }
+
+    ngAfterViewInit() {
+        this.initTitleMaxWidth();
     }
 
     isActiveRoute(route: string) {
@@ -49,6 +57,20 @@ export class DatasetViewComponent {
             queryParams[ShowVocUrlParams.hideDatasetName] = this.activatedRoute.snapshot.queryParams[ShowVocUrlParams.hideDatasetName];
         }
         this.router.navigate([route], { relativeTo: this.activatedRoute, queryParams: queryParams });
+    }
+
+    @HostListener('window:resize', ['$event'])
+    initTitleMaxWidth() {
+        let tabsWidth: number = 0;
+        let tabs: HTMLCollection = this.navtabsEl.nativeElement.children;
+        for (let i = 0; i < tabs.length; i++) {
+            let tab: Element = tabs.item(i);
+            if (tab instanceof HTMLElement) {
+                tabsWidth += tab.offsetWidth;
+            }
+        }
+        this.titleMaxWidth = this.navtabsEl.nativeElement.offsetWidth - tabsWidth - 40; //40 gives an additional margin
+        this.cdRef.detectChanges();
     }
 
 }