+++ ChangeLog.md
@@ -5,7 +5,7 @@ The full changelog of the platform includes also changes brought to the Semantic
 https://bitbucket.org/art-uniroma2/semantic-turkey/src/master/ChangeLog.txt
 
 # 2.2.2 (dd-mm-202y)
-  * 
+  * Added possibility to insert custom content into the home page blank section
 
 # 2.2.1 (19-10-2022)
   * Minor changes