+++ ChangeLog.md
@@ -7,6 +7,7 @@ https://bitbucket.org/art-uniroma2/semantic-turkey/src/master/ChangeLog.txt
 # 2.2.2 (dd-mm-202y)
   * Added possibility to insert custom content into the home page blank section
   * Enabled visualization of ordered collection members in ResourceView
+  * Fixed a bug that prevented to delete files uploaded in distributions panel
 
 # 2.2.1 (19-10-2022)
   * Minor changes