+++ ChangeLog.md
@@ -8,6 +8,7 @@ https://bitbucket.org/art-uniroma2/semantic-turkey/src/master/ChangeLog.txt
   * Added possibility to insert custom content into the home page blank section
   * Enabled visualization of ordered collection members in ResourceView
   * Fixed a bug that prevented to delete files uploaded in distributions panel
+  * Fixed minor UI issues
 
 # 2.2.1 (19-10-2022)
   * Minor changes